+++ test/configuration_test.rb
@@ -10,11 +10,13 @@ class ConfigurationTest < Minitest::Test
   def test_merge_updates_known_attributes
     @configuration.merge!(
       mount_path: "/terms",
+      app_name: "Harbor",
       require_scroll_to_end: "true",
       capture_request_provenance: "1"
     )
 
     assert_equal "/terms", @configuration.mount_path
+    assert_equal "Harbor", @configuration.app_name
     assert_equal true, @configuration.require_scroll_to_end
     assert_equal true, @configuration.capture_request_provenance
   end
@@ -44,6 +46,7 @@ class ConfigurationTest < Minitest::Test
     configuration = RecordingStudioTermsAndConditions::Configuration.new
 
     assert_equal "/recording_studio_terms_and_conditions", configuration.mount_path
+    assert_equal "", configuration.app_name
     assert_equal false, configuration.require_scroll_to_end
     assert_equal false, configuration.capture_request_provenance
     assert_instance_of RecordingStudio::Hooks, configuration.hooks
@@ -65,6 +68,7 @@ class ConfigurationTest < Minitest::Test
 
     assert_equal 2, result.fetch(:hooks_registered).fetch(:before_initialize)
     assert_equal 1, result.fetch(:hooks_registered).fetch(:after_service)
+    assert_equal "", result.fetch(:app_name)
     assert_equal false, result.fetch(:require_scroll_to_end)
     assert_equal false, result.fetch(:capture_request_provenance)
     refute result.key?(:required_categories)
@@ -77,5 +81,83 @@ class ConfigurationTest < Minitest::Test
     RecordingStudioTermsAndConditions.configure
 
     assert_kind_of RecordingStudioTermsAndConditions::Configuration, RecordingStudioTermsAndConditions.configuration
+  end
+
+  def test_app_name_returns_the_set_string
+    @configuration.app_name = "Harbor"
+
+    assert_equal "Harbor", @configuration.app_name
+  end
+
+  def test_app_name_blank_stays_empty_without_site_settings
+    without_site_settings do
+      @configuration.app_name = "   "
+
+      assert_equal "", @configuration.app_name
+    end
+  end
+
+  def test_app_name_falls_back_to_site_settings_name_for
+    with_site_settings_name("Harbor") do
+      @configuration.app_name = ""
+
+      assert_equal "Harbor", @configuration.app_name
+    end
+  end
+
+  def test_app_name_set_value_wins_over_site_settings
+    with_site_settings_name("Harbor") do
+      @configuration.app_name = "Terms Dummy"
+
+      assert_equal "Terms Dummy", @configuration.app_name
+    end
+  end
+
+  def test_app_name_blank_when_site_settings_has_no_name_for
+    with_site_settings_module do
+      @configuration.app_name = nil
+
+      assert_equal "", @configuration.app_name
+    end
+  end
+
+  private
+
+  def without_site_settings
+    stash_site_settings
+    yield
+  ensure
+    restore_site_settings
+  end
+
+  def with_site_settings_name(name)
+    stash_site_settings
+    mod = Module.new
+    mod.define_singleton_method(:name_for) { |*_args| name }
+    Object.const_set(:RecordingStudioSiteSettings, mod)
+    yield
+  ensure
+    restore_site_settings
+  end
+
+  def with_site_settings_module
+    stash_site_settings
+    Object.const_set(:RecordingStudioSiteSettings, Module.new)
+    yield
+  ensure
+    restore_site_settings
+  end
+
+  def stash_site_settings
+    @stashed_site_settings = Object.const_defined?(:RecordingStudioSiteSettings)
+    @site_settings_const = Object.const_get(:RecordingStudioSiteSettings) if @stashed_site_settings
+    Object.send(:remove_const, :RecordingStudioSiteSettings) if @stashed_site_settings
+  end
+
+  def restore_site_settings
+    if Object.const_defined?(:RecordingStudioSiteSettings)
+      Object.send(:remove_const, :RecordingStudioSiteSettings)
+    end
+    Object.const_set(:RecordingStudioSiteSettings, @site_settings_const) if @stashed_site_settings
   end
 end